+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.6.24] - 2023-08-12
+
 ### Added
 
 - Add array of transferable objects to the `execute()` method arguments.